moved test folders to proper package
